cleanup, some docs, temp fix for InputHandler tests
